requirements.txt
asgiref==3.2.3
Django==3.0
django-cors-headers==3.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==5.0.1
